+++ src/python/dxpy/scripts/dx.py
@@ -5430,6 +5430,10 @@ parser_update_project.add_argument('--external-upload-restricted', choices=["tru
 parser_update_project.add_argument('--database-ui-view-only', choices=["true", "false"],
                                    help="Whether the viewers on the project can access the database data directly")
 parser_update_project.add_argument('--bill-to', help="Update the user or org ID of the billing account", type=str)
+parser_update_project.add_argument('--https-app-isolated-browsing',
+                                   help="Whether all https access to jobs in this project should be wrapped in isolated browsing."
+                                        " If true, httpsApp-enabled executables must have httpsApp.shared_access set to 'NONE' to run in this project.",
+                                   choices=["true", "false"])
 allowed_executables_group = parser_update_project.add_mutually_exclusive_group()
 allowed_executables_group.add_argument('--allowed-executables', help='Executable ID(s) this project is allowed to run.  This operation overrides any existing list of executables.', type=str, nargs="+")
 allowed_executables_group.add_argument('--unset-allowed-executables', help='Removes any restriction to run executables as set by --allowed-executables', action='store_true')